+++ client/scripts/generate-static-route-entrypoints.js
@@ -11,6 +11,7 @@ const fallbackRoutes = [
   '/arenda-mikrofona-dlya-intervyu-moskva',
   '/arenda-kolonki-dlya-vecherinki-moskva',
   '/arenda-paroochistitelya-dlya-kuhni-plitki-vannoy-moskva',
+  '/arenda-kamery-dlya-puteshestviya-vloga-moskva',
   '/arenda-gopro-moskva',
   '/arenda-audiooborudovaniya-moskva',
   '/blog',